Initial Commit
On click of a choice buttons the next section of the story is displayed,
the the other unselected choice button is removed, and a "jump to" is added in the menu for the new section.
Upon completion of the story, an ending is created based off the previous choices. Based on the Bootstrap – Agency template.