Added Story Results divs
Added Story Result divs after each mini-story to both show the results of the user’s choice and to add a pause before the beginning of the next story. Adjusted functions to hide choice buttons that are already used and/or no longer needed. Adjusted functions using “innerHTML” to jQuery “.text”.